Ticket #761: FDO API: Use INT_MAX instead of LONG_MAX for FdoInt32. LONG_MAX == LLONG_MAX on Linux x64
git-svn-id: http://svn.osgeo.org/fdo/trunk@6927 d98eaee6-4212-0410-900c-dcc49bf69b99

+++ Fdo/UnitTest/SchemaTest.cpp
@@ -1402,7 +1402,7 @@ void SchemaTest::testDataConstraints ()
         1
     );
 
-    FdoInt32 int32List[5] = {10, 20, 30, LONG_MIN, LONG_MAX};
+    FdoInt32 int32List[5] = {10, 20, 30, INT_MIN, INT_MAX};
     AddListProperty(
         featClass,
         L"Int32Prop", 
@@ -1546,7 +1546,7 @@ void SchemaTest::testDataConstraints ()
         4
     );
 
-    FdoInt32 int32ListValues[10] = {9, 10, 11, 20, 0, 30, LONG_MIN + 1, LONG_MIN, LONG_MAX - 1, LONG_MAX};
+    FdoInt32 int32ListValues[10] = {9, 10, 11, 20, 0, 30, INT_MIN + 1, INT_MIN, INT_MAX - 1, INT_MAX};
     CheckConstraintProperty(
         featClass,
         L"Int32Prop",
